py/requirements.txt
alembic==1.1.0
attrs==20.1.0
boto3==1.14.48
botocore==1.17.48
click==7.1.2
docutils==0.15.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.8
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
